Switch to single integration with compat option
